backend/requirements.txt
# Build tools (install first)
wheel
setuptools

# Core framework - latest stable versions
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
# psycopg2 for alembic migrations (sync operations)
psycopg2-binary==2.9.10

# Redis
redis==5.2.1

# Monitoring
prometheus-client==0.21.1

# Security
bcrypt==4.2.1

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0